win32 tweaks
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1150 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/endpoints/mod_dingaling/mod_dingaling.c
@@ -123,7 +123,7 @@ struct private_object {
 	uint8_t codec_num;
 	switch_time_t last_cand;
 	char *stun_ip;
-	uint32_t stun_port;
+	uint16_t stun_port;
 };
 
 struct rfc2833_digit {
